Fix for newly introduced bug with dropping precision.  Restore some tests to String dates.  Make a new test with period tests and move tests there.  Clean up junk.
