+++ cqf-fhir-cr/src/test/java/org/opencds/cqf/fhir/cr/measure/r4/CollectData.java
@@ -6,7 +6,10 @@ import static org.opencds.cqf.fhir.test.Resources.getResourcePath;
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.parser.IParser;
 import java.nio.file.Paths;
+import java.time.LocalDate;
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.function.Supplier;
 import org.hl7.fhir.r4.model.IdType;
 import org.hl7.fhir.r4.model.MeasureReport;
@@ -124,6 +127,18 @@ public class CollectData {
 
         public When periodStart(ZonedDateTime periodStart) {
             this.periodStart = periodStart;
+            return this;
+        }
+
+        public When periodEnd(String periodEnd) {
+            this.periodEnd =
+                    LocalDate.parse(periodEnd, DateTimeFormatter.ISO_LOCAL_DATE).atStartOfDay(ZoneId.systemDefault());
+            return this;
+        }
+
+        public When periodStart(String periodStart) {
+            this.periodStart = LocalDate.parse(periodStart, DateTimeFormatter.ISO_LOCAL_DATE)
+                    .atStartOfDay(ZoneId.systemDefault());
             return this;
         }
